Use usize for truncate method in place of u64

+++ core/io/io_uring.rs
@@ -354,10 +354,10 @@ impl File for UringFile {
         Ok(self.file.metadata()?.len())
     }
 
-    fn truncate(&self, len: u64, c: Completion) -> Result<Arc<Completion>> {
+    fn truncate(&self, len: usize, c: Completion) -> Result<Arc<Completion>> {
         let mut io = self.io.borrow_mut();
         let truncate = with_fd!(self, |fd| {
-            io_uring::opcode::Ftruncate::new(fd, len)
+            io_uring::opcode::Ftruncate::new(fd, len as u64)
                 .build()
                 .user_data(io.ring.get_key())
         });
