+++ core/io/vfs.rs
@@ -166,7 +166,7 @@ impl File for VfsFileImpl {
         }
     }
 
-    fn truncate(&self, len: u64, c: Completion) -> Result<Arc<Completion>> {
+    fn truncate(&self, len: usize, c: Completion) -> Result<Arc<Completion>> {
         if self.vfs.is_null() {
             return Err(LimboError::ExtensionError("VFS is null".to_string()));
         }
